+++ Controls/_input/interface/IOnlyPositive.ts
@@ -6,7 +6,7 @@
 interface IOnlyPositive {
     readonly _options: {
         /**
-         * @name Controls/interface/IOnlyPositive#onlyPositive
+         * @name Controls/_input/interface/IOnlyPositive#onlyPositive
          * @cfg {Boolean} Определяет, можно ли вводить в поле только положительные числа.
          * @default false
          * @remark
@@ -19,22 +19,8 @@ interface IOnlyPositive {
          * <Controls.input:Number bind:value="_inputValue" onlyPositive="{{true}}"/>
          * </pre>
          */
-
-        /*
-         * @cfg {Boolean} Determines whether only positive numbers can be entered in the field.
-         * @default false
-         * @remark
-         * true - only positive numbers can be entered in the field.
-         * false - positive and negative numbers can be entered in the field.
-         * @example
-         * In this example you _inputValue in the control state will store only a positive number.
-         * <pre class="brush: html">
-         * <!-- WML -->
-         * <Controls.input:Number bind:value="_inputValue" onlyPositive="{{true}}"/>
-         * </pre>
-         */         
         onlyPositive: boolean;
-    }
+    };
 }
 
 export default IOnlyPositive;
