+++ Controls/interface.ts
@@ -75,3 +75,5 @@ export {Direction, IQueryParamsMeta, IQueryParams} from './_interface/IQueryPara
 export {default as IGrouping, IGroupingOptions} from './_interface/IGrouping';
 export {default as IPromiseSelectable, IPromiseSelectableOptions} from './_interface/IPromiseSelectable';
 export {default as IDraggable} from './_interface/IDraggable';
+export {default as IInputPlaceholder} from './_interface/IInputPlaceholder';
+export {default as IInputTag} from './_interface/IInputTag';
