+++ Controls/interface/IInputBase.ts
@@ -1,6 +1,5 @@
 import IPaste from "./IPaste";
-import IInputTag from "./IInputTag";
-import IInputPlaceholder from "./IInputPlaceholder";
+import {IInputPlaceholder, IInputTag} from 'Controls/interface';
 import ISelectableInput from "./ISelectableInput";
 import IInputTooltip from "./IInputTooltip";
 import IInputStyle from "./IInputStyle";
@@ -17,6 +16,7 @@ type IInputBase =
     & {
     readonly _options: {
         textAlign: 'left' | 'right' | 'center';
+        style: 'info' | 'invalid' | 'danger' | 'success' | 'warning' | 'primary'; //Todo нет такой опции больше
         autoComplete: 'on' | 'off' | 'username' | 'current-password';
     }
 };
