--- Controls/interface/IInputStyle.ts
@@ -1,7 +1,0 @@
-interface IInputStyle {
-    readonly _options: {
-        style: 'info' | 'invalid' | 'danger' | 'success' | 'warning' | 'primary';
-    };
-}
-
-export default IInputStyle;
